pyproject.toml
[project]
name = "llm-council"
version = "0.1.0"
description = "Your LLM Council"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.9.0",
    "langgraph>=0.4.0",
    "langchain-core>=0.3.0",
    "groq>=0.25.0",
    "PyPDF2>=3.0.0",
    "pytesseract>=0.3.10",
    "beautifulsoup4>=4.12.0",
    "duckduckgo-search>=7.0.0",
    "Pillow>=10.0.0",
    "python-multipart>=0.0.9",
]
